refactor: prepare patientAdmission to activity switch
